add algo located hotels

+++ HMS/HMSHotel.h
@@ -68,7 +68,7 @@
  @param stars the hotel stars
  @returns a newly initialized object
  */
-- (id)initWithParams:(NSString *)name :(NSString *)country :(NSString *)city :(NSString *) description :(NSInteger)stars :(float)latitude :(float)longitude;
+- (id)initWithParams:(NSString *)name :(NSString *)country :(NSString *)city :(NSString *) description :(NSInteger)stars :(double)latitude :(double)longitude;
 
 /**
  Add a room in this hotel
